requirements.txt
audiomentations==0.28.0
librosa==0.9.0
numpy==1.23.0
pandas==1.5.3
pydub==0.25.1
pytest==7.2.2
PyYAML==6.0
scikit_learn==1.2.2
soundfile==0.12.1
torch==1.13.1
xlsxwriter==3.0.9
tensorboard
ThreadPoolExecutor
